# executor: change outer result to list in index merge join