--- a/executor/index_lookup_merge_join.go
+++ b/executor/index_lookup_merge_join.go
@@ -89,8 +89,8 @@ type innerMergeCtx struct {
 }
 
 type lookUpMergeJoinTask struct {
-	outerResult   *chunk.Chunk
-	outerOrderIdx []int
+	outerResult   *chunk.List
+	outerOrderIdx []chunk.RowPtr
 
 	innerResult *chunk.Chunk
 	innerIter   chunk.Iterator
@@ -332,29 +332,35 @@ func (omw *outerMergeWorker) pushToChan(ctx context.Context, task *lookUpMergeJo
 func (omw *outerMergeWorker) buildTask(ctx context.Context) (*lookUpMergeJoinTask, error) {
 	task := &lookUpMergeJoinTask{
 		results:     make(chan *indexMergeJoinResult, numResChkHold),
-		outerResult: newFirstChunk(omw.executor),
+		outerResult: chunk.NewList(omw.rowTypes, omw.executor.base().initCap, omw.executor.base().maxChunkSize),
 	}
 	task.memTracker = memory.NewTracker(stringutil.MemoizeStr(func() string { return fmt.Sprintf("lookup join task %p", task) }), -1)
 	task.memTracker.AttachTo(omw.parentMemTracker)
 
 	omw.increaseBatchSize()
-	if omw.lookup.isOuterJoin { // if is outerJoin, push the requiredRows down
-		requiredRows := int(atomic.LoadInt64(&omw.lookup.requiredRows))
-		task.outerResult.SetRequiredRows(requiredRows, omw.maxBatchSize)
-	} else {
-		task.outerResult.SetRequiredRows(omw.batchSize, omw.maxBatchSize)
+	requiredRows := omw.batchSize
+	if omw.lookup.isOuterJoin {
+		requiredRows = int(atomic.LoadInt64(&omw.lookup.requiredRows))
+	}
+	if requiredRows <= 0 || requiredRows > omw.maxBatchSize {
+		requiredRows = omw.maxBatchSize
+	}
+	for requiredRows > 0 {
+		execChk := newFirstChunk(omw.executor)
+		err := Next(ctx, omw.executor, execChk)
+		if err != nil {
+			return task, err
+		}
+		if execChk.NumRows() == 0 {
+			break
+		}
+
+		task.outerResult.Add(execChk)
+		requiredRows -= execChk.NumRows()
+		task.memTracker.Consume(execChk.MemoryUsage())
 	}
 
-	task.memTracker.Consume(task.outerResult.MemoryUsage())
-	oldMemUsage := task.outerResult.MemoryUsage()
-	err := Next(ctx, omw.executor, task.outerResult)
-	if err != nil {
-		return task, err
-	}
-
-	newMemUsage := task.outerResult.MemoryUsage()
-	task.memTracker.Consume(newMemUsage - oldMemUsage)
-	if task.outerResult == nil || task.outerResult.NumRows() == 0 {
+	if task.outerResult.Len() == 0 {
 		return nil, nil
 	}
 
@@ -396,20 +402,26 @@ func (imw *innerMergeWorker) run(ctx context.Context, wg *sync.WaitGroup, cancel
 }
 
 func (imw *innerMergeWorker) handleTask(ctx context.Context, task *lookUpMergeJoinTask) (err error) {
-	numOuterRows := task.outerResult.NumRows()
+	numOuterRows := task.outerResult.Len()
 	var outerMatch []bool
 	if imw.outerMergeCtx.filter != nil {
 		outerMatch = make([]bool, numOuterRows)
 		task.memTracker.Consume(int64(cap(outerMatch)))
-		outerMatch, err = expression.VectorizedFilter(imw.ctx, imw.outerMergeCtx.filter, chunk.NewIterator4Chunk(task.outerResult), outerMatch)
+		outerMatch, err = expression.VectorizedFilterForList(imw.ctx, imw.outerMergeCtx.filter, task.outerResult, outerMatch)
 		if err != nil {
 			return err
 		}
 	}
-	task.outerOrderIdx = make([]int, 0, numOuterRows)
+	task.outerOrderIdx = make([]chunk.RowPtr, 0, numOuterRows)
+	totalChkSize, chkIdx, nowChkSize := task.outerResult.GetChunk(0).NumRows(), 0, 0
 	for i := 0; i < numOuterRows; i++ {
+		if i >= totalChkSize {
+			nowChkSize = totalChkSize
+			chkIdx++
+			totalChkSize += task.outerResult.GetChunk(chkIdx).NumRows()
+		}
 		if len(outerMatch) == 0 || outerMatch[i] {
-			task.outerOrderIdx = append(task.outerOrderIdx, i)
+			task.outerOrderIdx = append(task.outerOrderIdx, chunk.RowPtr{ChkIdx: uint32(chkIdx), RowIdx: uint32(i - nowChkSize)})
 		}
 	}
 	task.memTracker.Consume(int64(cap(task.outerOrderIdx)))
@@ -594,7 +606,7 @@ func (imw *innerMergeWorker) constructDatumLookupKeys(task *lookUpMergeJoinTask)
 	return dLookUpKeys, nil
 }
 
-func (imw *innerMergeWorker) constructDatumLookupKey(task *lookUpMergeJoinTask, rowIdx int) (*indexJoinLookUpContent, error) {
+func (imw *innerMergeWorker) constructDatumLookupKey(task *lookUpMergeJoinTask, rowIdx chunk.RowPtr) (*indexJoinLookUpContent, error) {
 	outerRow := task.outerResult.GetRow(rowIdx)
 	sc := imw.ctx.GetSessionVars().StmtCtx
 	keyLen := len(imw.keyCols)