--- a/expression/chunk_executor.go
+++ b/expression/chunk_executor.go
@@ -14,12 +14,13 @@
 package expression
 
 import (
+	"strconv"
+
 	"github.com/pingcap/parser/ast"
 	"github.com/pingcap/parser/mysql"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
-	"strconv"
 )
 
 // Vectorizable checks whether a list of expressions can employ vectorized execution.
@@ -331,6 +332,22 @@ func executeToString(ctx sessionctx.Context, expr Expression, fieldType *types.F
 		output.AppendString(colID, res)
 	}
 	return nil
+}
+
+// VectorizedFilterForList applies a list of filters to a List and
+// returns a bool slice, which indicates whether a row is passed the filters.
+// Filters is executed vectorized.
+func VectorizedFilterForList(ctx sessionctx.Context, filters []Expression, list *chunk.List, selected []bool) (_ []bool, err error) {
+	totalChkSize := 0
+	for i := 0; i < list.NumChunks(); i++ {
+		chk := list.GetChunk(i)
+		rows := chk.NumRows()
+		_, err = VectorizedFilter(ctx, filters, chunk.NewIterator4Chunk(chk), selected[totalChkSize:totalChkSize+rows])
+		if err != nil {
+			return selected, err
+		}
+	}
+	return selected, nil
 }
 
 // VectorizedFilter applies a list of filters to a Chunk and