--- a/expression/chunk_executor.go
+++ b/expression/chunk_executor.go
@@ -334,24 +334,6 @@ func executeToString(ctx sessionctx.Context, expr Expression, fieldType *types.F
 	return nil
 }
 
-// VectorizedFilterForList applies a list of filters to a List and
-// returns a bool slice, which indicates whether a row is passed the filters.
-// Filters is executed vectorized.
-func VectorizedFilterForList(ctx sessionctx.Context, filters []Expression, list *chunk.List, selected []bool) (_ []bool, err error) {
-	totalChkSize := 0
-	numChk := list.NumChunks()
-	for i := 0; i < numChk; i++ {
-		chk := list.GetChunk(i)
-		rows := chk.NumRows()
-		totalChkSize += rows
-		_, err = VectorizedFilter(ctx, filters, chunk.NewIterator4Chunk(chk), selected[totalChkSize:totalChkSize+rows])
-		if err != nil {
-			return selected, err
-		}
-	}
-	return selected, nil
-}
-
 // VectorizedFilter applies a list of filters to a Chunk and
 // returns a bool slice, which indicates whether a row is passed the filters.
 // Filters is executed vectorized.