<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
As the title says. The outer result is changed to list to avoid fetch little data in the outer side and the memory copy.

Reference issue: https://github.com/pingcap/tidb/issues/12498

### What is changed and how it works?
First, add a vectorized filter function for list. Then change some iteration logic.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test

Code changes

 - Has exported function/method change

Side effects

 - Increased code complexity