@lzmhhh123 2019-10-15
New benchmark compares with index hash join.
```bash
$ CGO_ENABLED=1 GO111MODULE=on go test -count 5 -benchmem -run=^$ github.com/pingcap/tidb/executor -bench BenchmarkIndexJoinExec
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/executor
BenchmarkIndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12                      8         223784819 ns/op        608316862 B/op    304097 allocs/op
BenchmarkIndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12                      8         258324156 ns/op        579590974 B/op    304086 allocs/op
BenchmarkIndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12                      7         294850408 ns/op        568235004 B/op    304058 allocs/op
BenchmarkIndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12                     12         296859596 ns/op        562721833 B/op    304040 allocs/op
BenchmarkIndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12                      9         287201640 ns/op        569117074 B/op    304046 allocs/op
BenchmarkIndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                      7        262968188 ns/op        568202419 B/op    304019 allocs/op
BenchmarkIndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                     7         293523160 ns/op        584710332 B/op    304062 allocs/op
BenchmarkIndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                     6         262443727 ns/op        591506274 B/op    304046 allocs/op
BenchmarkIndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                    13         330152695 ns/op        570100248 B/op    304022 allocs/op
BenchmarkIndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                     3         346654902 ns/op        639537576 B/op    304115 allocs/op
BenchmarkIndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                1        1431682940 ns/op        2482053168 B/op   307324 allocs/op
BenchmarkIndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                1        1138102730 ns/op        2482030664 B/op   307157 allocs/op
BenchmarkIndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                2        1388877594 ns/op        2482049436 B/op   307301 allocs/op
BenchmarkIndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                3        1574714213 ns/op        2482024701 B/op   307216 allocs/op
BenchmarkIndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                1        1211453381 ns/op        2482039104 B/op   307232 allocs/op
BenchmarkIndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                2        1071477132 ns/op        2472659020 B/op   208629 allocs/op
BenchmarkIndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                2        1199482944 ns/op        2472663360 B/op   208657 allocs/op
BenchmarkIndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                2        1249291263 ns/op        2472656300 B/op   208608 allocs/op
BenchmarkIndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                3        1540364218 ns/op        2472658824 B/op   208621 allocs/op
BenchmarkIndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                                1        1370140536 ns/op        2472659128 B/op   208627 allocs/op
PASS
ok      github.com/pingcap/tidb/executor        115.958s
```