@lzmhhh123 2019-10-18
Compare with the master version, hacked to chunk append like index join.

```bash
# liuzhuomin @ Charmings-Macbook-Pro in ~/.gvm/pkgsets/go1.13/global/src/github.com/pingcap/tidb on git:dev/index_merge_join_outerlist x [14:17:15]
$ benchstat old.txt new.txt
name                                                                                                                                                                      old time/op    new time/op    delta
IndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12     1.27s ±46%     0.26s ±20%  -79.32%  (p=0.008 n=5+5)
IndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                    1.26s ±45%     0.27s ±17%  -78.87%  (p=0.008 n=5+5)
IndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12               1.14s ±35%     1.17s ±32%     ~     (p=1.000 n=5+5)
IndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12               1.39s ±33%     1.19s ±26%     ~     (p=0.421 n=5+5)

name                                                                                                                                                                      old alloc/op   new alloc/op   delta
IndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12    3.38GB ± 0%    0.55GB ± 6%  -83.77%  (p=0.008 n=5+5)
IndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                   3.38GB ± 0%    0.59GB ± 6%  -82.67%  (p=0.008 n=5+5)
IndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12              2.48GB ± 0%    2.48GB ± 0%     ~     (p=1.000 n=5+5)
IndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12              2.47GB ± 0%    2.47GB ± 0%     ~     (p=0.056 n=5+5)

name                                                                                                                                                                      old allocs/op  new allocs/op  delta
IndexJoinExec/index_merge_join_need_outer_sort_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:true)-12      304k ± 0%      304k ± 0%     ~     (p=0.063 n=5+5)
IndexJoinExec/index_merge_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                     304k ± 0%      304k ± 0%   +0.06%  (p=0.008 n=5+5)
IndexJoinExec/index_inner_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                307k ± 0%      307k ± 0%     ~     (p=1.000 n=5+5)
IndexJoinExec/index_outer_hash_join_(outerRows:100000,_innerRows:102400,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-12                209k ± 0%      209k ± 0%     ~     (p=0.087 n=5+5)
```